+++ prisma/migrations/6_add_comm_reference_number_unique/migration.sql
@@ -1,0 +1,2 @@
+-- CreateIndex
+CREATE UNIQUE INDEX `communications_reference_number_key` ON `communications`(`reference_number`);