+++ src/modules/index-document/index-document.controller.ts
@@ -105,7 +105,7 @@ export class IndexDocumentController {
   ) {
     const isBlackAndWhite = color === 'bw';
     const isScanEffect = scanEffect === 'true';
-    return this.service.composeDocument(id, isBlackAndWhite, isScanEffect, dto?.imageIds ?? [], dto?.crops, dto?.watermark);
+    return this.service.composeDocument(id, isBlackAndWhite, isScanEffect, dto?.imageIds ?? [], dto?.crops, dto?.watermark, dto?.overrideImages);
   }
 
   @Delete(':id')